+++ modules/audio/examples/play_mp3.m
@@ -1,0 +1,17 @@
+
+%=============================================================================
+% Copyright (c) 2016-present Allan CORNET (Nelson)
+%=============================================================================
+% This file is part of the Nelson.
+%=============================================================================
+% LICENCE_BLOCK_BEGIN
+% SPDX-License-Identifier: LGPL-3.0-or-later
+% LICENCE_BLOCK_END
+%=============================================================================
+wav_audio = [modulepath('audio'), '/examples/opening.mp3'];
+[y, fs] = audioread(wav_audio);
+playObj = audioplayer(y, fs);
+playblocking(playObj)
+delete(playObj)
+clear playObj
+%=============================================================================